pyproject.toml
[project]
name = "tiktok-userbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "matplotlib>=3.10.1",
    "prophet>=1.1.6",
    "pyrogram>=2.0.106",
    "python-decouple>=3.8",
    "tgcrypto>=1.2.5",
]
